(SDK-144) Add option to run validate in parallel
This adds the `--parallel` flag to the `pdk validate` command. This
flag will cause the validations to run in their own threads.

This feature is currently hidden. It will not show up in the help
description, but is functional. Currently all PO files must be up-to-date and
contain no obsolete or out-of-date strings to work with the parallel validation.